requirements.txt
numpy>=1.24.0
pyvista>=0.43.0
trimesh>=4.0.0
PyYAML>=6.0
numpy-stl>=3.0.0
scipy>=1.10.0
matplotlib>=3.7.0
networkx>=2.8.0
scikit-image>=0.20.0
manifold3d>=3.0.0
